+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Fixed
+
+- Using a namespaced column as a property of a specification failed when using `Sign::in`
+
 ## 2.5.0 - 2023-08-06
 
 ### Added